+++ plugins/tv-dashboard/src/components/deck/index.ts
@@ -1,0 +1,7 @@
+export { ComunicadoElementInspector } from "./ComunicadoElementInspector";
+export { ComunicadoSlideBackgroundPanel } from "./ComunicadoSlideBackgroundPanel";
+export { DeckActionRow } from "./DeckActionRow";
+export { DeckField } from "./DeckField";
+export { DeckInspectorLayout } from "./DeckInspectorLayout";
+export { DeckPropertySection } from "./DeckPropertySection";
+export { DeckRibbonShell } from "./DeckRibbonShell";